+++ hourai/extensions/validation/rejectors.py
@@ -149,3 +149,15 @@ class BannedUsernameRejector(Validator):
 
     def _normalize(self, val):
         return " ".join(val.casefold().split())
+
+
+class LockdownRejector(Validator):
+    """A malice level validator that rejects all users if the guild has been put
+    into lockdown.
+    """
+
+    async def validate_member(self, ctx):
+        guild_state = ctx.bot.guild_states[ctx.guild.id]
+        if guild_state.is_locked_down:
+            ctx.add_rejection_reason(
+                'Lockdown enabled. All new joins must be manually verified.')